+++ htslib/hfile.h
@@ -248,7 +248,7 @@ hwrite(hFILE *fp, const void *buffer, size_t nbytes)
     extern int hfile_set_blksize(hFILE *fp, size_t bufsiz);
 
     if(!fp->mobile){
-        if (fp->limit - fp->begin < nbytes){
+        if ((size_t)(fp->limit - fp->begin) < nbytes){
             hfile_set_blksize(fp, fp->limit - fp->buffer + nbytes);
             fp->end = fp->limit;
         }
